requirements.txt

streamlit
tensorflow-cpu
tensorflow-hub
pillow
requests
numpy